Finished general feel of the front end based on wireframe. Added files to begin filtering access to pages.
