+++ app/controllers/serviceLearning/__init__.py
@@ -8,3 +8,6 @@ serviceLearning_bp = Blueprint(
 )
 
 from app.controllers.serviceLearning import routes
+from app.controllers.serviceLearning import slcProposal
+from app.controllers.serviceLearning import slcGuidelines
+from app.controllers.serviceLearning import slcQuestionnaire
