+++ app/controllers/serviceLearning/slcProposal.py
@@ -1,0 +1,9 @@
+from flask import request, render_template, g, abort
+from app.controllers.serviceLearning import serviceLearning_bp
+
+
+@serviceLearning_bp.route('/slcProposal')
+def slcProposal():
+    """This page allows faculties to create service learning proposal"""
+
+    return render_template('serviceLearning/slcProposal.html')
