+++ app/templates/serviceLearning/slcGuidelines.html
@@ -11,7 +11,7 @@
   <h1 class="title text-center pb-4 fs-3"> Service-Learning Course Designation Active Learning Experience (ALE) Requirements met through Service-Learning </h1>
   <h2 class="subtitle text-center pb-2 fs-5"> CRITERIA AND PROPOSAL FORM</h2>
   <p>
-    SERVICE-LEARNING at Berea College is aneducational experience based upon a collaborative partnership between Collegeand community. Learning through service enables students to apply academic knowledge and critical-thinking skills to meet genuine community needs. Through reflection and assessment, students gain deeper understanding of course content and the importance of civic engagement.
+    SERVICE-LEARNING at Berea College is an educational experience based upon a collaborative partnership between College and community. Learning through service enables students to apply academic knowledge and critical-thinking skills to meet genuine community needs. Through reflection and assessment, students gain deeper understanding of course content and the importance of civic engagement.
   </p>
   <p>
     All Designated Service-Learning Courses will meet the Active Learning Experience (ALE) Requirement.
@@ -33,7 +33,7 @@
         Demonstrate deeper understanding of specific course learning objectives appropriate to the experience.
       </li>
       <li>
-        Reflect critically on multiple facets of their experience, on the connections between theory and practice, andbetween learning inside and outside the classroom
+        Reflect critically on multiple facets of their experience, on the connections between theory and practice, and between learning inside and outside the classroom
       </li>
       <li>
         Faculty will identify one or more course learning objectives that students will address through their service work in the community.
