+++ app/templates/serviceLearning/slcProposal.html
@@ -10,7 +10,7 @@
 
 <h1 class="text-center pb-2 fs-2"> New Proposal For Service Learning Course Designation </h1>
 
-<p class="text-center "> Please review the <a href="/slcGuidelines">Service Learning course guidelines</a> before completing this form </p>
+<p class="text-center fw-bold"> Please review the <a href="/slcGuidelines">Service Learning course guidelines</a> before completing this form </p>
 <div class="container-fluid">
   <div class="row">
     <div class="p-2 col-md-4">
@@ -49,7 +49,7 @@
   </div>
   <div class="col" align="right">
     <label for="Continue">(2/3)</label>
-    <a href="#" id="Continue" type="button" class="btn btn-success col">Continue</a>
+    <a href="/slcQuestionnaire" id="Continue" type="button" class="btn btn-success col">Continue</a>
   </div>
 
 </div>
