+++ app/templates/serviceLearning/slcQuestionnaire.html
@@ -1,0 +1,46 @@
+{% extends "base.html" %}
+{% block scripts %}
+
+  {{super()}}
+  <style media="screen"></style>
+
+{% endblock %}
+
+{% block app_content %}
+
+<h1 class="text-center">New Designated Service Learning Course</h1>
+<p class="text-center fw-bold"> Please review the <a href="/slcGuidelines">Service Learning course guidelines</a> before completing this form </p>
+<div class="pt-4">
+  <label>1.Indicate the specific course learning objective(s) that will be addressed by the service-learning component</label>
+  <textarea class="form-control" aria-label="With textarea"></textarea>
+  <label>2. Describe the method of structured, critical reflection you will use in order to connect the service
+            activity or activities to the learning objective(s) of the course.</label>
+  <textarea class="form-control" aria-label="With textarea"></textarea>
+  <label>3. Describe the service-learning activity or activities.  Indicate the approximate number of hours or
+            days of service experience. There is no formal requirement, but the activity must be substantial and
+            sustained throughout the semester. </label>
+  <textarea class="form-control" aria-label="With textarea"></textarea>
+  <label>4. Identify the community partner(s) with whom you will collaborate in order to structure this
+            service-learning experience.  How will the community partner(s) be involved in planning, implementation,
+            and evaluation?</label>
+  <textarea class="form-control" aria-label="With textarea"></textarea>
+  <label>5. Describe the final product to be completed by the students.  How will this benefit the community? </label>
+  <textarea class="form-control" aria-label="With textarea"></textarea>
+  <label>6.Indicate how you will assess student learning related to the identified course objective(s) that will
+          be enhanced by the service-learning experience. Identify what percentage of the final grade will be based
+          on the assessment of reflection, the final product, and other forms of assessment related to the identified
+          course objective(s).</label>
+  <textarea class="form-control" aria-label="With textarea"></textarea>
+</div>
+<div class="row align-items-end pt-5">
+  <div class="col">
+    <a href="/slcProposal" type="button" class="btn btn-primary col">Previous</a>
+  </div>
+  <div class="col" align="right">
+    <label for="Continue">(3/3)</label>
+    <a href="/slcQuestionnaire" id="Continue" type="button" class="btn btn-success col">Submit</a>
+  </div>
+
+</div>
+
+{% endblock %}
